back/requirements.txt
Django==5.2.1
django-filter==25.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
drf-nested-routers==0.94.1
drf-spectacular==0.28.0
python-dotenv==1.1.0
psycopg2-binary==2.9.10

pytest==8.3.5
pytest-django==4.11.1
pytest-xdist==3.6.1
pytest-cov==6.1.1
